requirements.txt
tqdm
dask
distributed>=2.3.0
flake8
fsspec
isodate
matplotlib
mypy
mypy_extensions
networkx
numpy
pandas==0.24.2|>0.25.2
pytest
scikit-learn==0.21.1|>0.21.2
scipy
pytables
tensorboardX
pytorch
torchvision
typeguard
fastparquet>=0.3.2
bokeh
jupyter
jupyter_contrib_nbextensions
xgboost
